+++ docs/release-notes.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+* feat: lipsync-server APIルーター + main.py. PR [#25](https://github.com/Rx-K8/AnotherMe/pull/25) by [@Rx-K8](https://github.com/Rx-K8).
 * feat: lipsync-server Service層 + MockProvider + Factory. PR [#24](https://github.com/Rx-K8/AnotherMe/pull/24) by [@Rx-K8](https://github.com/Rx-K8).
 * feat: lipsync-serverバリデーター（audio/media）追加. PR [#22](https://github.com/Rx-K8/AnotherMe/pull/22) by [@Rx-K8](https://github.com/Rx-K8).
 * feat: lipsync-serverのプロジェクト骨格とインターフェース定義. PR [#19](https://github.com/Rx-K8/AnotherMe/pull/19) by [@Rx-K8](https://github.com/Rx-K8).